[Init] Added Initial Project Structure
To design the initial structure of the project, I followed how other
projects tend to structure themselves. I found Chapter 14.3 very useful
for this. I treated the overall git repo as a Cargo workspace and added
the viewer as the main executable and the parser as a library. This
should allow allow for separation of the two pieces of code.