[Parse] Added Support for the Stratix-IV Arch
The Stratix-IV arch is a model of a real intel FPGA. It is quite a bit
more complicated than the other architectures we have been using for
testing.

This architecture was actually auto-generated by a python tool (for at
least large parts of it), so it really exercises the documentation
available. A bit of the VTR documentation will need to be updated.